リファクタリング実施: miku-grep の実装とテストを責務別に分割する
## 概要

`miku-grep` の Node CLI 実装を責務別のモジュールに分割し、巨大化していたテストを機能別に整理します。

## 変更内容

- `src/main.ts` に集約されていた処理を分離
  - help text: `src/help.ts`
  - glob / path matching: `src/glob.ts`
  - request validation: `src/validation.ts`
  - result / summary / diagnostics helper: `src/result.ts`
  - traversal / file search: `src/search.ts`
- 型定義の境界を整理
  - 公開 request / result 型を `src/public-types.ts` に分離
  - 内部作業型を `src/internal-types.ts` に分離
  - `src/types.ts` は互換用 re-export facade として維持
- `test/miku-grep-cli.test.ts` を削除し、機能別テストへ分割
  - CLI meta / stdio contract
  - validation
  - content search
  - filename / both search
  - encoding / diagnostics
  - output limits
  - 共通 fixture / helper
- `scripts/bundle-smoke.mjs` の source archive 検査対象を新しいファイル構成に更新
- `TODO.md` と CLI spec の Development Notes を現状の構成に更新
- JSON Schema はMVPでは導入しない判断を `TODO.md` に記録

## 確認

- `npm test` 成功
- `npm run build` 成功
- `npm run smoke` 成功
- `npm run smoke:bundle` 成功
- `npm_config_cache=.npm-cache npm pack --dry-run` 成功
